Inventory implemented
Still lacks functionality, but the basis of an inventory is there. Today the entire day was spent on trial and error, trying to make a good inventory system, so, due to the time constraint, I decided to follow Code Monkey's video on inventory. It's basic, it's not scalable or modular, but it gets the job done. If there's any time left, I will remake this system in a better way, but, unfortunately, for now, it will have to do.